Remove generic paramater on bloom filter
Having a generic parameter `T` on `CountingBloomFilter<T>` together with
a method `CountingBloomFilter::insert(&mut self, item: &T)` is very
restrictive, because it does not allow inserting an item `U` where
`T: Borrow<U>` (as is done in Rust's HashMap implementation). Taking the
example of `String`, one couldn't insert a `&str`. On the other hand, if
one were to set `CountingBloomFilter<&'a str>`, lifetimes can become a
problem is the filter is supposed live longer than the items that are
inserted into it (a very reasonable assumption). This necessitates that
`T` is an owned quantity - or at least a quantity with `T: 'static`.

At the same time, `T: Borrow<U>` is still overly restrictive if all one
cares about is that two types `T` and `U` behave the same when hashed.
As an example, given a `CountingBloomFilter<Vec<String>>`, there exists
an `impl Borrow<[T]> for Vec<T>` such that one can get a `&[String]`
from a `Vec<String>`. However, one might want to insert a `&[&'a str]`
into the filter, which however cannot be easily constructed via `Borrow`
or `Deref`.

The easiest solution is thus to allow any type `T` to be hashed into the
bloom filter and have the user ensure that they insert the correct
items.