Fix incorrect calculation of bucket indices
Insertion, containment, and count estimation logic did not correctly
calculate the index of a bucket from the calculated hashes. It did
not take into account the number of bits that belong to a bucket, so
that it took the bitslice at range `i..i+bits` as the bits belonging to
bucket `i`, which is clearly wrong (and worse, leads to overlapping
buckets!). The correct bit indices are at `j = bits * i`, i.e.
`(bits * i)..(bits * i)+bits`.